+++ src/outputs/sdl_out.ml
@@ -99,7 +99,7 @@ class output ~infallible ~on_start ~on_stop ~autostart ~kind source =
 let () =
   let kind = Lang.video_yuva420p in
   let k = Lang.kind_type_of_kind_format kind in
-  Lang.add_operator "output.sdl" ~active:true
+  Lang.add_operator "output.sdl"
     (Output.proto @ [("", Lang.source_t k, None, None)])
     ~return_t:k ~category:Lang.Output ~descr:"Display a video using SDL."
     (fun p ->